split out functions to respecive files and some cleanup
