+++ ld_script.txt
@@ -76,8 +76,10 @@ SECTIONS {
         src/file_system.o(.text);
         src/main.o(.text);
         asm/code_800B5F0.o(.text);
+        src/code_800B5F0.o(.text);
         src/music.o(.text);
         src/code_800C9CC.o(.text);
+        src/bg_control.o(.text);
         src/cpu.o(.text);
         src/flash.o(.text);
         src/code_800D090.o(.text);
